♻️ refactor: extract EmailAccountsPanel from CredentialsPage
Splits the 1260-line god component into focused modules.
EmailAccountsPanel owns all email config state, handlers, and JSX.
CredentialsPage now composes three independent panels.
