+++ web/src/features/credentials/CredentialsPage.tsx
@@ -6,7 +6,6 @@ import {
   disconnectOAuth as disconnectOAuthRequest,
   getOAuthConnected,
   getOAuthProviderConfig,
-  getVaultEntry,
   listOAuthProviders,
   listVaultEntries,
   pollOAuthFlow,
@@ -23,6 +22,7 @@ import { useI18n } from "@/lib/i18n";
 import { useToast, ToastContainer } from "@/hooks/use-toast";
 import { meQueryOptions } from "@/lib/queries/me";
 import { SettingsPageHeader } from "@/features/settings/SettingsPageHeader";
+import { EmailAccountsPanel } from "@/features/credentials/EmailAccountsPanel";
 import { siGithub, siX } from "simple-icons";
 
 const SIMPLE_ICON_PATHS: Record<string, string> = {
@@ -52,49 +52,6 @@ function simpleIconPath(name: string) {
   return SIMPLE_ICON_PATHS[name.toLowerCase()];
 }
 
-interface EmailAccount {
-  imap_host: string;
-  imap_port: number;
-  imap_tls: string;
-  smtp_host: string;
-  smtp_port: number;
-  smtp_tls: string;
-  username: string;
-  password?: string;
-  from: string;
-}
-
-interface EmailConfig {
-  default: string;
-  accounts: Record<string, EmailAccount>;
-}
-
-interface EmailFormValues {
-  name: string;
-  imapHost: string;
-  imapPort: string;
-  imapTls: string;
-  smtpHost: string;
-  smtpPort: string;
-  smtpTls: string;
-  username: string;
-  from: string;
-  password: string;
-}
-
-const INITIAL_EMAIL_FORM: EmailFormValues = {
-  name: "",
-  imapHost: "",
-  imapPort: "993",
-  imapTls: "ssl",
-  smtpHost: "",
-  smtpPort: "587",
-  smtpTls: "starttls",
-  username: "",
-  from: "",
-  password: "",
-};
-
 export function CredentialsPage() {
   const { t } = useI18n();
   const { data: me } = useQuery(meQueryOptions);
@@ -105,15 +62,6 @@ export function CredentialsPage() {
   const [vaultSaving, setVaultSaving] = useState(false);
   const [newSecretName, setNewSecretName] = useState("");
   const [newSecretValue, setNewSecretValue] = useState("");
-
-  const [emailConfig, setEmailConfig] = useState<EmailConfig>({ default: "", accounts: {} });
-  const [emailConfigLoading, setEmailConfigLoading] = useState(false);
-  const [emailConfigSaving, setEmailConfigSaving] = useState(false);
-  const [isEmailFormOpen, setIsEmailFormOpen] = useState(false);
-  const [editingAccountName, setEditingAccountName] = useState<string | null>(null);
-
-  const [emailFormValues, setEmailFormValues] = useState<EmailFormValues>(INITIAL_EMAIL_FORM);
-  const [emailFormErrors, setEmailFormErrors] = useState<Record<string, string>>({});
 
   const [oauthProviders, setOauthProviders] = useState<OAuthProvider[]>([]);
   const [oauthStatus, setOauthStatus] = useState<
@@ -218,207 +166,9 @@ export function CredentialsPage() {
     }
   }, []);
 
-  const loadEmailConfig = useCallback(async () => {
-    setEmailConfigLoading(true);
-    try {
-      const { data } = await getVaultEntry({
-        path: { name: "EMAIL_CONFIG" },
-        throwOnError: true,
-      });
-      if (data && data.value) {
-        const parsed = JSON.parse(data.value) as EmailConfig;
-        if (!parsed.accounts) {
-          parsed.accounts = {};
-        }
-        setEmailConfig(parsed);
-      }
-    } catch {
-      // 404/NotFound or error means no email config yet
-      setEmailConfig({ default: "", accounts: {} });
-    } finally {
-      setEmailConfigLoading(false);
-    }
-  }, []);
-
-  const handleSetDefaultEmail = async (name: string) => {
-    const updated = {
-      ...emailConfig,
-      default: name,
-    };
-    setEmailConfigSaving(true);
-    try {
-      await setVaultEntry({
-        path: { name: "EMAIL_CONFIG" },
-        body: { value: JSON.stringify(updated) },
-        throwOnError: true,
-      });
-      showToast("Default account updated");
-      await loadEmailConfig();
-    } catch (e) {
-      showToast(e instanceof Error ? e.message : "Failed to update default account", "error");
-    } finally {
-      setEmailConfigSaving(false);
-    }
-  };
-
-  const handleDeleteEmail = async (name: string) => {
-    if (!window.confirm(`Delete email account "${name}"?`)) return;
-
-    const updatedAccounts = { ...emailConfig.accounts };
-    delete updatedAccounts[name];
-
-    let nextDefault = emailConfig.default;
-    if (nextDefault === name) {
-      const keys = Object.keys(updatedAccounts);
-      nextDefault = keys.length > 0 ? keys[0] : "";
-    }
-
-    const updated = {
-      default: nextDefault,
-      accounts: updatedAccounts,
-    };
-
-    setEmailConfigSaving(true);
-    try {
-      if (Object.keys(updatedAccounts).length === 0) {
-        await deleteVaultEntryRequest({ path: { name: "EMAIL_CONFIG" }, throwOnError: true });
-        setEmailConfig({ default: "", accounts: {} });
-      } else {
-        await setVaultEntry({
-          path: { name: "EMAIL_CONFIG" },
-          body: { value: JSON.stringify(updated) },
-          throwOnError: true,
-        });
-        await loadEmailConfig();
-      }
-      showToast("Email account deleted");
-    } catch (e) {
-      showToast(e instanceof Error ? e.message : "Failed to delete account", "error");
-    } finally {
-      setEmailConfigSaving(false);
-    }
-  };
-
-  const handleEditEmail = (name: string) => {
-    const acct = emailConfig.accounts[name];
-    if (!acct) return;
-
-    setEmailFormValues({
-      name: name,
-      imapHost: acct.imap_host || "",
-      imapPort: String(acct.imap_port || 993),
-      imapTls: acct.imap_tls || "ssl",
-      smtpHost: acct.smtp_host || "",
-      smtpPort: String(acct.smtp_port || 587),
-      smtpTls: acct.smtp_tls || "starttls",
-      username: acct.username || "",
-      from: acct.from || "",
-      password: "",
-    });
-    setEditingAccountName(name);
-    setIsEmailFormOpen(true);
-    setEmailFormErrors({});
-  };
-
-  const handleAddEmail = () => {
-    setEmailFormValues(INITIAL_EMAIL_FORM);
-    setEditingAccountName(null);
-    setIsEmailFormOpen(true);
-    setEmailFormErrors({});
-  };
-
-  const handleSaveEmail = async () => {
-    const errors: Record<string, string> = {};
-    if (!editingAccountName) {
-      if (!emailFormValues.name) {
-        errors.name = "Account name is required";
-      } else if (!/^[a-z][a-z0-9_]{0,31}$/.test(emailFormValues.name)) {
-        errors.name =
-          "Must start with a lowercase letter and contain only lowercase letters, numbers, and underscores (max 32 chars)";
-      } else if (emailConfig.accounts && emailFormValues.name in emailConfig.accounts) {
-        errors.name = "Account name already exists";
-      }
-    }
-
-    if (!emailFormValues.imapHost) errors.imapHost = "IMAP host is required";
-    if (!emailFormValues.smtpHost) errors.smtpHost = "SMTP host is required";
-    if (!emailFormValues.username) errors.username = "Username is required";
-    if (!emailFormValues.from) errors.from = "From address is required";
-
-    const imapPortNum = parseInt(emailFormValues.imapPort, 10);
-    if (isNaN(imapPortNum) || imapPortNum <= 0 || imapPortNum > 65535) {
-      errors.imapPort = "Invalid port number";
-    }
-
-    const smtpPortNum = parseInt(emailFormValues.smtpPort, 10);
-    if (isNaN(smtpPortNum) || smtpPortNum <= 0 || smtpPortNum > 65535) {
-      errors.smtpPort = "Invalid port number";
-    }
-
-    if (!editingAccountName && !emailFormValues.password) {
-      errors.password = "Password/App Password is required";
-    }
-
-    if (Object.keys(errors).length > 0) {
-      setEmailFormErrors(errors);
-      return;
-    }
-
-    setEmailConfigSaving(true);
-    try {
-      const name = editingAccountName || emailFormValues.name;
-      const existingAcct = emailConfig.accounts?.[name];
-
-      const newAccount: EmailAccount = {
-        imap_host: emailFormValues.imapHost,
-        imap_port: imapPortNum,
-        imap_tls: emailFormValues.imapTls,
-        smtp_host: emailFormValues.smtpHost,
-        smtp_port: smtpPortNum,
-        smtp_tls: emailFormValues.smtpTls,
-        username: emailFormValues.username,
-        from: emailFormValues.from,
-        password:
-          emailFormValues.password !== ""
-            ? emailFormValues.password
-            : (existingAcct?.password ?? ""),
-      };
-
-      const updatedAccounts = {
-        ...emailConfig.accounts,
-        [name]: newAccount,
-      };
-
-      let nextDefault = emailConfig.default;
-      if (!nextDefault || Object.keys(emailConfig.accounts || {}).length === 0) {
-        nextDefault = name;
-      }
-
-      const updated = {
-        default: nextDefault,
-        accounts: updatedAccounts,
-      };
-
-      await setVaultEntry({
-        path: { name: "EMAIL_CONFIG" },
-        body: { value: JSON.stringify(updated) },
-        throwOnError: true,
-      });
-
-      showToast(editingAccountName ? "Email account updated" : "Email account added");
-      setIsEmailFormOpen(false);
-      await loadEmailConfig();
-    } catch (e) {
-      showToast(e instanceof Error ? e.message : "Failed to save account", "error");
-    } finally {
-      setEmailConfigSaving(false);
-    }
-  };
-
   useEffect(() => {
     const init = async () => {
       await loadVaultEntries();
-      await loadEmailConfig();
       const providers = await loadOAuthProviders();
       await Promise.all([
         ...providers.map((p) => checkOAuthConnected(p.provider)),
@@ -431,14 +181,7 @@ export function CredentialsPage() {
         pollAbortRef.current[key] = true;
       }
     };
-  }, [
-    loadVaultEntries,
-    loadEmailConfig,
-    loadOAuthProviders,
-    checkOAuthConnected,
-    loadProviderConfig,
-    isAdmin,
-  ]);
+  }, [loadVaultEntries, loadOAuthProviders, checkOAuthConnected, loadProviderConfig, isAdmin]);
 
   const addVaultEntry = useCallback(async () => {
     if (!newSecretName) {
@@ -861,324 +604,6 @@ export function CredentialsPage() {
     </div>
   );
 
-  const hasAccounts = emailConfig.accounts && Object.keys(emailConfig.accounts).length > 0;
-
-  const emailDetail = (
-    <div className="space-y-6">
-      {emailConfigLoading && <p className="text-sm text-muted-foreground">Loading…</p>}
-
-      {hasAccounts && (
-        <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-          {Object.entries(emailConfig.accounts).map(([name, acct]) => {
-            const isDefault = emailConfig.default === name;
-            return (
-              <div
-                key={name}
-                className="min-w-0 rounded-2xl border border-border/40 bg-card/45 backdrop-blur-md p-5 shadow-2xs flex flex-col justify-between"
-              >
-                <div>
-                  <div className="flex min-w-0 items-start justify-between gap-3">
-                    <div className="flex min-w-0 items-center gap-3">
-                      <svg
-                        viewBox="0 0 24 24"
-                        fill="none"
-                        stroke="currentColor"
-                        strokeWidth="2"
-                        className="size-4 text-muted-foreground/80"
-                      >
-                        <rect width="20" height="16" x="2" y="4" rx="2" />
-                        <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-                      </svg>
-                      <span className="min-w-0 truncate text-sm font-medium font-mono">{name}</span>
-                    </div>
-                    {isDefault && <Badge variant="success">Default</Badge>}
-                  </div>
-
-                  <div className="mt-4 space-y-2 text-xs text-muted-foreground">
-                    <div>
-                      <span className="text-foreground font-medium">From:</span> {acct.from}
-                    </div>
-                    <div>
-                      <span className="text-foreground font-medium">Username:</span> {acct.username}
-                    </div>
-                    <div>
-                      <span className="text-foreground font-medium">IMAP:</span> {acct.imap_host}:
-                      {acct.imap_port || 993} ({acct.imap_tls})
-                    </div>
-                    <div>
-                      <span className="text-foreground font-medium">SMTP:</span> {acct.smtp_host}:
-                      {acct.smtp_port || 587} ({acct.smtp_tls})
-                    </div>
-                  </div>
-                </div>
-
-                <div className="mt-5 flex items-center justify-end gap-2 border-t border-border/20 pt-3">
-                  {!isDefault && (
-                    <Button
-                      size="xs"
-                      variant="ghost"
-                      onClick={() => handleSetDefaultEmail(name)}
-                      loading={emailConfigSaving}
-                    >
-                      Set Default
-                    </Button>
-                  )}
-                  <Button size="xs" variant="outline" onClick={() => handleEditEmail(name)}>
-                    Edit
-                  </Button>
-                  <Button
-                    size="xs"
-                    variant="destructive-outline"
-                    className="text-destructive hover:bg-destructive/10"
-                    onClick={() => handleDeleteEmail(name)}
-                    loading={emailConfigSaving}
-                  >
-                    Delete
-                  </Button>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      )}
-
-      {!hasAccounts && !emailConfigLoading && (
-        <p className="text-sm text-muted-foreground py-4 text-center">
-          No email accounts configured yet.
-        </p>
-      )}
-
-      {!isEmailFormOpen ? (
-        <div className="flex justify-end">
-          <Button size="sm" onClick={handleAddEmail}>
-            Add Email Account
-          </Button>
-        </div>
-      ) : (
-        <div className="rounded-2xl border border-border/40 bg-card/45 backdrop-blur-md p-6 shadow-2xs space-y-6">
-          <div className="flex items-center justify-between">
-            <h3 className="text-sm font-semibold text-foreground/90">
-              {editingAccountName
-                ? `Edit Email Account: ${editingAccountName}`
-                : "Add Email Account"}
-            </h3>
-            <Button size="xs" variant="ghost" onClick={() => setIsEmailFormOpen(false)}>
-              Cancel
-            </Button>
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            {!editingAccountName && (
-              <div className="space-y-1.5 md:col-span-2">
-                <label className="text-xs font-medium text-muted-foreground">Account Name</label>
-                <Input
-                  type="text"
-                  value={emailFormValues.name}
-                  onChange={(e) => setEmailFormValues({ ...emailFormValues, name: e.target.value })}
-                  placeholder="e.g. personal, work"
-                  autoComplete="off"
-                  nativeInput
-                />
-                {emailFormErrors.name && (
-                  <p className="text-xs text-destructive">{emailFormErrors.name}</p>
-                )}
-              </div>
-            )}
-
-            {/* IMAP Config */}
-            <div className="space-y-4 rounded-xl border border-border/25 bg-muted/10 p-4">
-              <h4 className="text-xs font-semibold text-muted-foreground uppercase tracking-wider">
-                IMAP (Incoming)
-              </h4>
-
-              <div className="space-y-1.5">
-                <label className="text-xs font-medium text-muted-foreground">IMAP Host</label>
-                <Input
-                  type="text"
-                  value={emailFormValues.imapHost}
-                  onChange={(e) =>
-                    setEmailFormValues({ ...emailFormValues, imapHost: e.target.value })
-                  }
-                  placeholder="imap.example.com"
-                  autoComplete="off"
-                  nativeInput
-                />
-                {emailFormErrors.imapHost && (
-                  <p className="text-xs text-destructive">{emailFormErrors.imapHost}</p>
-                )}
-              </div>
-
-              <div className="grid grid-cols-2 gap-3">
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">Port</label>
-                  <Input
-                    type="text"
-                    value={emailFormValues.imapPort}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, imapPort: e.target.value })
-                    }
-                    placeholder="993"
-                    autoComplete="off"
-                    nativeInput
-                  />
-                  {emailFormErrors.imapPort && (
-                    <p className="text-xs text-destructive">{emailFormErrors.imapPort}</p>
-                  )}
-                </div>
-
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">TLS Mode</label>
-                  <select
-                    value={emailFormValues.imapTls}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, imapTls: e.target.value })
-                    }
-                    className="w-full rounded-lg border border-input bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-ring h-9"
-                  >
-                    <option value="ssl">SSL / TLS (recommended)</option>
-                    <option value="starttls">STARTTLS</option>
-                    <option value="none">None</option>
-                  </select>
-                </div>
-              </div>
-            </div>
-
-            {/* SMTP Config */}
-            <div className="space-y-4 rounded-xl border border-border/25 bg-muted/10 p-4">
-              <h4 className="text-xs font-semibold text-muted-foreground uppercase tracking-wider">
-                SMTP (Outgoing)
-              </h4>
-
-              <div className="space-y-1.5">
-                <label className="text-xs font-medium text-muted-foreground">SMTP Host</label>
-                <Input
-                  type="text"
-                  value={emailFormValues.smtpHost}
-                  onChange={(e) =>
-                    setEmailFormValues({ ...emailFormValues, smtpHost: e.target.value })
-                  }
-                  placeholder="smtp.example.com"
-                  autoComplete="off"
-                  nativeInput
-                />
-                {emailFormErrors.smtpHost && (
-                  <p className="text-xs text-destructive">{emailFormErrors.smtpHost}</p>
-                )}
-              </div>
-
-              <div className="grid grid-cols-2 gap-3">
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">Port</label>
-                  <Input
-                    type="text"
-                    value={emailFormValues.smtpPort}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, smtpPort: e.target.value })
-                    }
-                    placeholder="587"
-                    autoComplete="off"
-                    nativeInput
-                  />
-                  {emailFormErrors.smtpPort && (
-                    <p className="text-xs text-destructive">{emailFormErrors.smtpPort}</p>
-                  )}
-                </div>
-
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">TLS Mode</label>
-                  <select
-                    value={emailFormValues.smtpTls}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, smtpTls: e.target.value })
-                    }
-                    className="w-full rounded-lg border border-input bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-ring h-9"
-                  >
-                    <option value="starttls">STARTTLS (recommended)</option>
-                    <option value="ssl">SSL / TLS</option>
-                    <option value="none">None</option>
-                  </select>
-                </div>
-              </div>
-            </div>
-
-            {/* Credentials / Auth */}
-            <div className="space-y-4 rounded-xl border border-border/25 bg-muted/10 p-4 md:col-span-2">
-              <h4 className="text-xs font-semibold text-muted-foreground uppercase tracking-wider">
-                Account Credentials
-              </h4>
-
-              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">Username</label>
-                  <Input
-                    type="text"
-                    value={emailFormValues.username}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, username: e.target.value })
-                    }
-                    placeholder="user@example.com"
-                    autoComplete="off"
-                    nativeInput
-                  />
-                  {emailFormErrors.username && (
-                    <p className="text-xs text-destructive">{emailFormErrors.username}</p>
-                  )}
-                </div>
-
-                <div className="space-y-1.5">
-                  <label className="text-xs font-medium text-muted-foreground">From Address</label>
-                  <Input
-                    type="text"
-                    value={emailFormValues.from}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, from: e.target.value })
-                    }
-                    placeholder="Name <user@example.com> or user@example.com"
-                    autoComplete="off"
-                    nativeInput
-                  />
-                  {emailFormErrors.from && (
-                    <p className="text-xs text-destructive">{emailFormErrors.from}</p>
-                  )}
-                </div>
-
-                <div className="space-y-1.5 md:col-span-2">
-                  <label className="text-xs font-medium text-muted-foreground">
-                    Password / App Password
-                  </label>
-                  <Input
-                    type="password"
-                    value={emailFormValues.password}
-                    onChange={(e) =>
-                      setEmailFormValues({ ...emailFormValues, password: e.target.value })
-                    }
-                    placeholder={
-                      editingAccountName ? "Keep existing password (leave blank)" : "password"
-                    }
-                    autoComplete="new-password"
-                    nativeInput
-                  />
-                  {emailFormErrors.password && (
-                    <p className="text-xs text-destructive">{emailFormErrors.password}</p>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-
-          <div className="flex justify-end gap-3 pt-2">
-            <Button size="sm" variant="ghost" onClick={() => setIsEmailFormOpen(false)}>
-              Cancel
-            </Button>
-            <Button size="sm" loading={emailConfigSaving} onClick={handleSaveEmail}>
-              Save Account
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-
   return (
     <div className="h-full overflow-y-auto bg-background">
       <div className="mx-auto max-w-3xl p-6 sm:p-8 lg:p-10 space-y-8">
@@ -1231,7 +656,7 @@ export function CredentialsPage() {
                 Email Accounts
               </h4>
             </div>
-            <div className="space-y-4">{emailDetail}</div>
+            <EmailAccountsPanel showToast={showToast} />
           </div>
 
           {/* Vault Secrets Section */}
